feat: Update Document model to include link field
This commit updates the Document model in the prisma/schema.prisma file to include a new field called "link". The "link" field will store a string value representing the link associated with the document. This change allows for the storage and retrieval of document links in the database, enhancing the functionality of the application.